changed the logic for adding of script
